switch to kusto sdk (#280)
* Initial Commit

* WIP

* WIP2

* Working KustoSDK

* Remove unused code

* small nit pick fix
